test(checker): split JSX union callback coverage
